[Dashboard Agent] inline dashboard app mount integration
Move dashboard attachment mount wiring into the dashboard app integration flow, remove the temporary on-attachment mount wrapper, and keep the latest origin sync adjustments together with that cleanup.

Made-with: Cursor
